# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Legacy-to-Salesforce Migration Platform API
  version: "1.4.0"
  description: |
    REST API for the Legacy-to-Salesforce Migration Platform. Provides full lifecycle
    management of migration jobs including creation, monitoring, control operations,
    validation, and reporting.

    ## Authentication

    All endpoints require authentication. Two schemes are supported:

    1. **OAuth 2.0 Client Credentials** — Recommended for service-to-service integrations
       and automated pipelines. Obtain a token from the token endpoint using client credentials.

    2. **API Key** — Suitable for read-only monitoring integrations. Pass the API key
       in the `X-API-Key` header. API keys are scoped to a single tenant and cannot
       perform write operations.

    ## Rate Limiting

    API endpoints are rate-limited per tenant:
    - GET endpoints: 300 req/min
    - POST/mutating endpoints: 60 req/min

    The following headers are returned on every response:
    - `X-RateLimit-Limit` — Maximum requests per minute for this endpoint
    - `X-RateLimit-Remaining` — Remaining requests in the current window
    - `X-RateLimit-Reset` — UTC epoch timestamp when the limit resets

    ## Versioning

    The API is versioned via URL path (`/api/v1/`). Breaking changes result in a new version.
    Deprecated endpoints are supported for 12 months after deprecation notice.

    ## Idempotency

...

  contact:
    name: Migration Platform Engineering
...
  license:
...
servers:
...
security:
  - OAuth2ClientCredentials:
      - migrations:read
      - migrations:write
components:
  securitySchemes:
    OAuth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.migration-platform.internal/oauth/token
          scopes:
            migrations:read: Read migration jobs, reports, and validation results
            migrations:write: Create and control migration jobs
            migrations:admin: Administrative operations — tenant management, configuration
            metrics:read: Read platform metrics
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Tenant-scoped API key for read-only monitoring integrations.
        Issued per tenant via the platform admin UI. Cannot perform write operations.
  parameters:
    MigrationIdParam:
      name: migrationId
      in: path
      required: true
      schema:
        type: string
      description: Migration job ID
      example: mig-20251201-a3f8b2c1
  schemas:
...
    MigrationStatus:
      type: string
      enum:
        - DRAFT
        - VALIDATING
        - READY
        - STARTING
        - RUNNING
        - PAUSED
        - COMPLETING
        - COMPLETED
        - FAILED
        - ROLLING_BACK
        - ROLLED_BACK
        - CANCELLED
      description: |
        Migration job state machine transitions:
        DRAFT → VALIDATING → READY → STARTING → RUNNING → COMPLETING → COMPLETED
        RUNNING → PAUSED → RUNNING (via resume)
        RUNNING|PAUSED|COMPLETED|FAILED → ROLLING_BACK → ROLLED_BACK
        RUNNING → FAILED
...
    CreateMigrationRequest:
      type: object
      required: [name, tenant_id, source, target]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: "Acme Corp - Oracle EBS to Salesforce - Phase 1"
        description:
          type: string
          maxLength: 2000
          example: "Migrating Customer and Contact data from Oracle EBS R12 to Salesforce Sales Cloud"
        tenant_id:
          type: string
          description: Client tenant identifier — must match an existing provisioned tenant
          example: ent-acme-corp
        source:
          $ref: '#/components/schemas/MigrationSourceConfig'
        target:
          $ref: '#/components/schemas/MigrationTargetConfig'
        schedule:
          type: object
          properties:
            start_at:
              type: string
              format: date-time
              nullable: true
              description: Scheduled start time. Null means manual start via the start endpoint.
              example: "2025-12-01T22:00:00Z"
            maintenance_windows:
              type: array
              description: |
                Daily time windows during which loading to Salesforce is permitted.
                Extraction and transformation run continuously; loading pauses outside windows.
              items:
                type: object
                properties:
                  start_utc:
                    type: string
                    description: Window start time (HH:MM UTC)
                    example: "22:00"
                  end_utc:
                    type: string
                    description: Window end time (HH:MM UTC)
                    example: "06:00"
        options:
          type: object
          properties:
            error_threshold_pct:
              type: number
              format: double
              minimum: 0
              maximum: 100
              default: 2.0
              description: Auto-pause if DLQ error rate exceeds this percentage of total records
            enable_ai_triage:
              type: boolean
              default: true
              description: |
                Enable AI-assisted DLQ triage (requires Data Privacy Review approval;
                PII is masked before any AI API calls)
            notification_emails:
              type: array
              items:
                type: string
                format: email
              description: Email addresses for completion and failure notifications
            tags:
              type: object
              additionalProperties:
                type: string
              description: Arbitrary key-value metadata for filtering and reporting
              example:
                project: crm-modernization-q4-2025
                change_ticket: CHG-4821
    Migration:
      type: object
      required: [migration_id, name, tenant_id, status, created_at, updated_at]
      properties:
        migration_id:
          type: string
          example: mig-20251201-a3f8b2c1
        name:
          type: string
          example: "Acme Corp - Oracle EBS to Salesforce - Phase 1"
        description:
          type: string
          nullable: true
        tenant_id:
          type: string
          example: ent-acme-corp
        status:
          $ref: '#/components/schemas/MigrationStatus'
        created_at:
          type: string
          format: date-time
          example: "2025-12-01T20:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-12-01T22:14:37Z"
        started_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-12-01T22:00:03Z"
        completed_at:
          type: string
          format: date-time
          nullable: true
        estimated_completion:
          type: string
          format: date-time
          nullable: true
          example: "2025-12-02T14:30:00Z"
        phases:
          type: array
          items:
            $ref: '#/components/schemas/MigrationPhase'
        totals:
          type: object
          properties:
            records_expected:
              type: integer
              format: int64
              example: 2847391
            records_extracted:
              type: integer
              format: int64
              example: 2847391
            records_transformed:
              type: integer
              format: int64
              example: 2789044
            records_loaded:
              type: integer
              format: int64
              example: 2023847
            records_rejected:
              type: integer
              format: int64
              example: 58347
            records_in_dlq:
              type: integer
              format: int64
              example: 847
            success_rate_pct:
              type: number
              format: double
              example: 97.85
        source:
          $ref: '#/components/schemas/MigrationSourceConfig'
        target:
          $ref: '#/components/schemas/MigrationTargetConfig'
        created_by:
          type: string
          format: email
          example: data-team@acme.com
        _links:
          type: object
          properties:
            self:
              type: string
              format: uri
              example: /api/v1/migrations/mig-20251201-a3f8b2c1
            report:
              type: string
              format: uri
              example: /api/v1/migrations/mig-20251201-a3f8b2c1/report
            validation:
              type: string
              format: uri
              example: /api/v1/migrations/mig-20251201-a3f8b2c1/validation
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: VALIDATION_ERROR
            message: "Request validation failed"
            request_id: "a3f8b2c1-dead-beef-cafe-000000000001"
            timestamp: "2025-12-01T22:00:00.000Z"
            details:
              - field: source.entities
                message: "At least one source entity is required"
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: UNAUTHORIZED
            message: "Valid authentication credentials required"
            request_id: "b4e9c3d2-dead-beef-cafe-000000000002"
            timestamp: "2025-12-01T22:00:00.000Z"
    Forbidden:
      description: Authenticated but not authorized for this operation or tenant
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: FORBIDDEN
            message: "scope 'migrations:write' required for this operation"
            request_id: "c5f0d4e3-dead-beef-cafe-000000000003"
            timestamp: "2025-12-01T22:00:00.000Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: NOT_FOUND
            message: "Migration job mig-99999999-notfound not found"
            request_id: "d6a1e5f4-dead-beef-cafe-000000000004"
            timestamp: "2025-12-01T22:00:00.000Z"
    Conflict:
      description: Operation conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: INVALID_STATE_TRANSITION
            message: "Cannot pause a migration in COMPLETED state"
            request_id: "e7b2f6a5-dead-beef-cafe-000000000005"
            timestamp: "2025-12-01T22:00:00.000Z"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Number of seconds to wait before retrying
          example: 12
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: RATE_LIMIT_EXCEEDED
            message: "Rate limit of 60 requests/minute exceeded for POST endpoints"
            request_id: "f8c3a7b6-dead-beef-cafe-000000000006"
            timestamp: "2025-12-01T22:00:00.000Z"
paths:
  /migrations:
    post:
      tags: [migrations]
      operationId: createMigration
      summary: Create a new migration job
      description: |
        Creates a new migration job and runs pre-migration validation checks asynchronously.

        The job is created in `VALIDATING` state and transitions to:
        - `READY` — all pre-checks passed; awaiting manual start or scheduled start time
        - `FAILED` — a critical pre-check failed (connectivity, missing external ID field, etc.)

        Pre-checks performed automatically:
        - Source system connectivity and read permissions
        - Salesforce org connectivity and Bulk API permissions
        - Transformation rule set existence and approval status
        - Salesforce picklist mapping completeness
        - External ID field existence and API name verification
        - Kubernetes ResourceQuota headroom for tenant namespace
        - Vault credential availability for all required secrets

        Poll `GET /api/v1/migrations/{migrationId}` to track state transition out of VALIDATING.
        Typical pre-check duration: 30–120 seconds.
      security:
        - OAuth2ClientCredentials: [migrations:write]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
          description: Idempotency key to prevent duplicate job creation on retry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMigrationRequest'
            examples:
              oracle_ebs_migration:
                summary: Oracle EBS to Salesforce — full example
                value:
                  name: "Acme Corp - Oracle EBS to Salesforce - Phase 1"
                  description: "Migrating Customer and Contact data from Oracle EBS R12 to Salesforce Sales Cloud"
                  tenant_id: ent-acme-corp
                  source:
                    system_type: ORACLE_EBS
                    connection_name: oracle-ebs-prod-r2
                    entities:
                      - entity_name: HZ_PARTIES
                        rule_set_id: oracle-ebs-to-sf-account-v3
                        filter_expression: "PARTY_TYPE = 'ORGANIZATION' AND STATUS = 'A'"
                        estimated_records: 1500000
                      - entity_name: HZ_CONTACTS
                        rule_set_id: oracle-ebs-to-sf-contact-v2
                        estimated_records: 2847391
                  target:
                    salesforce_connection_name: acme-corp-sf-prod
                    api_version: "62.0"
                    bulk_api_concurrency: 10
                    objects:
                      - object_name: Account
                        operation: UPSERT
                        external_id_field: Legacy_ID__c
                      - object_name: Contact
                        operation: UPSERT
                        external_id_field: Legacy_ID__c
                  schedule:
                    start_at: "2025-12-01T22:00:00Z"
                    maintenance_windows:
                      - start_utc: "22:00"
                        end_utc: "06:00"
                  options:
                    error_threshold_pct: 2.0
                    enable_ai_triage: true
                    notification_emails:
                      - admin@acme.com
                      - data-team@acme.com
                    tags:
                      project: crm-modernization-q4-2025
                      change_ticket: CHG-4821
                      migrated_by: data-team@acme.com
      responses:
        '202':
          description: Migration job created; pre-validation running in background
          headers:
            Location:
              schema:
                type: string
              description: URL to retrieve migration status
              example: /api/v1/migrations/mig-20251201-a3f8b2c1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Migration'
              example:
                migration_id: mig-20251201-a3f8b2c1
                name: "Acme Corp - Oracle EBS to Salesforce - Phase 1"
                tenant_id: ent-acme-corp
                status: VALIDATING
                created_at: "2025-12-01T20:00:00Z"
                updated_at: "2025-12-01T20:00:01Z"
                started_at: null
                estimated_completion: null
                _links:
                  self: /api/v1/migrations/mig-20251201-a3f8b2c1
                  report: /api/v1/migrations/mig-20251201-a3f8b2c1/report
                  validation: /api/v1/migrations/mig-20251201-a3f8b2c1/validation
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Idempotency key conflict — duplicate request detected within 24h window
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /migrations/{migrationId}:
    get:
      tags: [migrations]
      operationId: getMigration
      summary: Get migration job status and progress
      description: |
        Returns current status, phase-level progress, and record count statistics
        for a migration job.

        Recommended polling interval during active migration: 30 seconds.
        When migration reaches COMPLETED or FAILED, no further polling is needed.
      security:
        - OAuth2ClientCredentials: [migrations:read]
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/MigrationIdParam'
      responses:
        '200':
          description: Migration job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Migration'
              example:
                migration_id: mig-20251201-a3f8b2c1
                name: "Acme Corp - Oracle EBS to Salesforce - Phase 1"
                tenant_id: ent-acme-corp
                status: RUNNING
                created_at: "2025-12-01T20:00:00Z"
                updated_at: "2025-12-01T22:14:37Z"
                started_at: "2025-12-01T22:00:03Z"
                completed_at: null
                estimated_completion: "2025-12-02T14:30:00Z"
                phases:
                  - name: EXTRACTION
                    status: COMPLETED
                    records_total: 2847391
                    records_processed: 2847391
                    records_failed: 0
                    throughput_per_minute: 0
                    started_at: "2025-12-01T22:00:05Z"
                    completed_at: "2025-12-01T23:47:12Z"
                    estimated_completion: null
                  - name: TRANSFORMATION
                    status: RUNNING
                    records_total: 2847391
                    records_processed: 2789044
                    records_failed: 58347
                    throughput_per_minute: 48750.5
                    started_at: "2025-12-01T22:00:47Z"
                    completed_at: null
                    estimated_completion: "2025-12-01T22:16:00Z"
                  - name: LOADING
                    status: RUNNING
                    records_total: 2789044
                    records_processed: 2023847
                    records_failed: 8471
                    throughput_per_minute: 4212.3
                    started_at: "2025-12-01T22:01:22Z"
                    completed_at: null
                    estimated_completion: "2025-12-02T14:30:00Z"
                totals:
                  records_expected: 2847391
                  records_extracted: 2847391
                  records_transformed: 2789044
                  records_loaded: 2023847
                  records_rejected: 58347
                  records_in_dlq: 847
                  success_rate_pct: 97.85
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...